requirements.txt
click==8.1.3
GitPython==3.1.31
graphviz==0.20.1
PyYAML==6.0
requests==2.28.2
tqdm==4.65.0
python-hcl2==4.3.0
ipaddr==2.2.0
debugpy==1.5.1